# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Outpost VPN API
  description: |
    Open-source WireGuard VPN & Zero-Trust Network Access platform.
    All features available under Apache 2.0 — no enterprise paywalls.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Romashqua Labs
    url: https://github.com/romashqua/outpost
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://vpn.example.com
    description: Production
...
paths:
...
  /api/v1/networks:
    get:
      operationId: listNetworks
      summary: List all VPN networks
      tags: [Networks]
      responses:
        "200":
          description: Network list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Network"

    post:
      operationId: createNetwork
      summary: Create a new VPN network
      tags: [Networks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNetworkRequest"
      responses:
        "201":
          description: Network created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Network already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/networks/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceID"

    get:
      operationId: getNetwork
      summary: Get network by ID
      tags: [Networks]
      responses:
        "200":
          description: Network found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
        "404":
          description: Network not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      operationId: updateNetwork
      summary: Update a network
      tags: [Networks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNetworkRequest"
      responses:
        "200":
          description: Network updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
        "404":
          description: Network not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      operationId: deleteNetwork
      summary: Delete a network
      tags: [Networks]
      responses:
        "204":
          description: Network deleted
        "404":
          description: Network not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Devices ─────────────────────────────────────────────────────────
...
  /api/v1/gateways:
    get:
      operationId: listGateways
      summary: List all WireGuard gateways
      tags: [Gateways]
      responses:
        "200":
          description: Gateway list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gateway"

    post:
      operationId: createGateway
      summary: Create a new gateway (auto-generates WG keypair)
      tags: [Gateways]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGatewayRequest"
      responses:
        "201":
          description: Gateway created with credentials (token & private_key returned only once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayWithCredentials"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/gateways/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceID"

    get:
      operationId: getGateway
      summary: Get gateway by ID
      tags: [Gateways]
      responses:
        "200":
          description: Gateway found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gateway"
        "404":
          description: Gateway not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      operationId: deleteGateway
      summary: Delete a gateway
      tags: [Gateways]
      responses:
        "204":
          description: Gateway deleted
        "404":
          description: Gateway not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Analytics ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via POST /api/v1/auth/login
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: something went wrong

    # ── Auth ────────────────────────────────────────────────────────
...
    Network:
      type: object
      required: [id, name, address, port, keepalive, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
          description: CIDR notation
          example: 10.10.0.0/24
        dns:
          type: array
          items:
            type: string
          example: ["1.1.1.1", "8.8.8.8"]
        port:
          type: integer
          example: 51820
        keepalive:
          type: integer
          description: PersistentKeepalive interval in seconds
          example: 25
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateNetworkRequest:
      type: object
      required: [name, address]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
          description: CIDR notation
          example: 10.10.0.0/24
        dns:
          type: array
          items:
            type: string
        port:
          type: integer
          default: 51820
        keepalive:
          type: integer
          default: 25
    UpdateNetworkRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        dns:
          type: array
          items:
            type: string
        port:
          type: integer
        keepalive:
          type: integer
        is_active:
          type: boolean

    # ── Devices ─────────────────────────────────────────────────────
...
    Gateway:
      type: object
      required: [id, network_id, name, wireguard_pubkey, endpoint, is_active, priority, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        network_id:
          type: string
          format: uuid
        name:
          type: string
        wireguard_pubkey:
          type: string
          description: WireGuard public key (base64)
        endpoint:
          type: string
          description: "host:port for WireGuard"
          example: vpn.example.com:51820
        is_active:
          type: boolean
        priority:
          type: integer
          description: Failover priority (lower = preferred)
        last_seen:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GatewayWithCredentials:
      description: Returned only on gateway creation — credentials are never retrievable again
      allOf:
        - $ref: "#/components/schemas/Gateway"
        - type: object
          required: [token, private_key]
          properties:
            token:
              type: string
              description: Authentication token for the gateway agent
            private_key:
              type: string
              description: WireGuard private key (base64) — store securely
    CreateGatewayRequest:
      type: object
      required: [name, network_id, endpoint]
      properties:
        name:
          type: string
          minLength: 1
        network_id:
          type: string
          format: uuid
        endpoint:
          type: string
          description: "host:port for WireGuard"
          example: gw1.example.com:51820

    # ── Analytics ───────────────────────────────────────────────────
...
security:
  - BearerAuth: []